+++ src/Header/Parameter.php
@@ -3,12 +3,50 @@ declare(strict_types = 1);
 
 namespace Innmind\Http\Header;
 
+use Innmind\Immutable\Str;
+
 /**
  * @psalm-immutable
  */
-interface Parameter
+final class Parameter
 {
-    public function name(): string;
-    public function value(): string;
-    public function toString(): string;
+    private string $name;
+    private string $value;
+    private string $string;
+
+    public function __construct(string $name, string $value)
+    {
+        $value = Str::of($value)->trim();
+
+        if ($value->matches("/[ \t]/")) {
+            $value = $value
+                ->trim('"')
+                ->append('"')
+                ->prepend('"');
+        }
+
+        $this->name = $name;
+        $this->value = $value->toString();
+        $this->string = \sprintf(
+            '%s%s%s',
+            $this->name,
+            \strlen($this->value) > 0 ? '=' : '',
+            \strlen($this->value) > 0 ? $this->value : '',
+        );
+    }
+
+    public function name(): string
+    {
+        return $this->name;
+    }
+
+    public function value(): string
+    {
+        return $this->value;
+    }
+
+    public function toString(): string
+    {
+        return $this->string;
+    }
 }