+++ src/Header/Parameter/Quality.php
@@ -3,6 +3,7 @@ declare(strict_types = 1);
 
 namespace Innmind\Http\Header\Parameter;
 
+use Innmind\Http\Header\Parameter;
 use Innmind\Immutable\Str;
 
 /**