+++ CHANGELOG.md
@@ -75,6 +75,7 @@
 - `Innmind\Http\Header\Parameter\Quality` no longer implements `Innmind\Http\Header\Parameter`
 - `Innmind\Http\Header\Parameter\Quality::of()` now returns `self`
 - `Innmind\Http\Header\ContentType\Boundary` no longer implements `Innmind\Http\Header\Parameter`
+- `Innmind\Http\Header\Parameter` is now a final class
 
 ### Removed